Removal of obsolete Invocation page and related utilities
